requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
attrs==24.2.0
customtkinter==5.2.2
darkdetect==0.8.0
discord-py==2.4.0
frozenlist==1.5.0
idna==3.10
keyboard==0.13.5
mouse==0.7.1
mss==10.0.0
multidict==6.1.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
python-dotenv==1.0.1
yarl==1.17.1
